Fix TTL rounding to match Redis behavior (#1602)
* Fix TTL rounding to match Redis behavior (#1389)

Redis rounds TTL seconds to the nearest integer (equivalent to (ms+500)/1000),
while Garnet was truncating (flooring) via a direct (long) cast. This caused
SET+EXPIRE 10+TTL to return 9 instead of 10 when called immediately.

Fix: Use Math.Round(MidpointRounding.AwayFromZero) in both
SecondsFromDiffUtcNowTicks and MillisecondsFromDiffUtcNowTicks in
ConvertUtils.cs.

Added TtlRoundingTest to verify string and object keys return the expected
TTL value immediately after EXPIRE.


* updates

* updates

* increase test timing tolerance

* update

---------

+++ test/Garnet.test/RespEtagTests.cs
@@ -1648,11 +1648,11 @@ namespace Garnet.test
             var expire = 2;
 
             var ttl = db.Execute("TTL", key);
-            ClassicAssert.AreEqual(-2, (int)ttl);
+            ClassicAssert.AreEqual(-2, (long)ttl);
 
             db.Execute("SET", [key, val, "WITHETAG"]);
             ttl = db.Execute("TTL", key);
-            ClassicAssert.AreEqual(-1, (int)ttl);
+            ClassicAssert.AreEqual(-1, (long)ttl);
 
             db.KeyExpire(key, TimeSpan.FromSeconds(expire));
 
@@ -2096,12 +2096,12 @@ namespace Garnet.test
             var expireTimeInMilliseconds = 3000;
 
             var pttl = db.Execute("PTTL", key);
-            ClassicAssert.AreEqual(-2, (int)pttl);
+            ClassicAssert.AreEqual(-2, (long)pttl);
 
             db.Execute("SET", [key, val, "WITHETAG"]);
 
             pttl = db.Execute("PTTL", key);
-            ClassicAssert.AreEqual(-1, (int)pttl);
+            ClassicAssert.AreEqual(-1, (long)pttl);
 
             db.KeyExpire(key, TimeSpan.FromMilliseconds(expireTimeInMilliseconds));
